solve @media problem
